[feat] 연령별 추천 SQL 추가   -테스트 코드 추가 필요

+++ src/main/java/letsnona/nonabackend/global/security/service/MemberServiceImpl.java
@@ -81,7 +81,7 @@ public class MemberServiceImpl implements MemberService {
 
         for (CountOfProductSellListAndPriceDTO list : countOfProductSellList
         ) {
-            totalPrice += (list.getPrice().intValue() * list.getCnt().intValue());
+            totalPrice += (list.getPrice() * list.getCnt().intValue());
         }
 
         sellProductRatioDTO.setTrading(trading);
